requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
langchain-chroma
chromadb
tiktoken
streamlit-extras
google-generativeai
python-dotenv

